refactor: extract block primitives into primitives/block.h
Lift the pure block/index value types (CBlockHeader, CBlock, CBlockIndex,
CDiskBlockIndex, CBlockLocator, BlockHasher, the GRC::MRCFees/MintSummary
reports, LOCKTIME_THRESHOLD, FutureDrift, GetTargetSpacing) out of the
~1200-line main.h god-header into a new, main.h-free primitives/block.h,
mirroring Bitcoin Core's primitives/block.h.

The eight inline methods that reference chain-state globals (pindexBest,
mapBlockIndex, pindexGenesisBlock, fUseFastIndex, the genesis hashes) are
relocated out-of-line into primitives/block.cpp so the header stays free of
main.h. Their EXCLUSIVE_LOCKS_REQUIRED(cs_main) annotations remain on the
in-class declarations.

consensus/merkle.h and consensus/tx_verify.h now include primitives/block.h
instead of main.h, removing the layering inversion where low-level consensus
headers dragged in the entire top-of-stack include graph. tx_verify.cpp gains
a direct main.h include for the chain-state globals it still uses. main.h keeps
functional change.
