+++ src/consensus/tx_verify.cpp
@@ -7,6 +7,7 @@
 #include <node/blockstorage.h>
 #include <primitives/transaction.h>
 #include "txdb.h"
+#include "main.h"
 
 bool IsFinalTx(const CTransaction &tx, int nBlockHeight, int64_t nBlockTime) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
 {
