+++ src/main.h
@@ -17,6 +17,7 @@
 #include "gridcoin/contract/contract.h"
 #include "gridcoin/cpid.h"
 #include "primitives/transaction.h"
+#include "primitives/block.h"
 #include "sync.h"
 #include "script.h"
 #include "scrypt.h"
@@ -47,8 +48,6 @@ class MRC;
 typedef std::optional<Claim> ClaimOption;
 }
 
-/** Threshold for nLockTime: below this value it is interpreted as block number, otherwise as UNIX timestamp. */
-static const unsigned int LOCKTIME_THRESHOLD = 500000000; // Tue Nov  5 00:53:20 1985 UTC
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -64,13 +63,6 @@ static const uint256 hashGenesisBlockRegTest = uint256S("0x4692b9564c585f76f41e0
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 
-inline int64_t FutureDrift(int64_t nTime, int nHeight) { return nTime + 20 * 60; }
-inline unsigned int GetTargetSpacing(int nHeight) { return IsProtocolV2(nHeight) ? 90 : 60; }
-
-struct BlockHasher
-{
-    size_t operator()(const uint256& hash) const { return hash.GetUint64(0); }
-};
 
 typedef std::unordered_map<uint256, CBlockIndex*, BlockHasher> BlockMap;
 
@@ -265,906 +257,6 @@ public:
 
 
 
-/** Nodes collect new transactions into a block, hash them into a hash tree,
- * and scan through nonce values to make the block's hash satisfy proof-of-work
- * requirements.  When they solve the proof-of-work, they broadcast the block
- * to everyone and the block is added to the block chain.  The first transaction
- * in the block is a special one that creates a new coin owned by the creator
- * of the block.
- *
- * Blocks are appended to blk0001.dat files on disk.  Their location on disk
- * is indexed by CBlockIndex objects in memory.
- */
-class CBlockHeader
-{
-public:
-    static const int32_t CURRENT_VERSION = 14;
-
-    // header
-    int32_t nVersion;
-    uint256 hashPrevBlock;
-    uint256 hashMerkleRoot;
-    uint32_t nTime;
-    uint32_t nBits;
-    uint32_t nNonce;
-
-    CBlockHeader()
-    {
-        SetNull();
-    }
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action)
-    {
-        READWRITE(nVersion);
-        READWRITE(hashPrevBlock);
-        READWRITE(hashMerkleRoot);
-        READWRITE(nTime);
-        READWRITE(nBits);
-
-        // Besides early blocks, Gridcoin uses Proof-of-Stake for consensus,
-        // so we don't need the nonce field. Don't serialize it after blocks
-        // version 11 and later:
-        //
-        if (nVersion <= 10) {
-            READWRITE(nNonce);
-        }
-    }
-
-    void SetNull()
-    {
-        nVersion = CURRENT_VERSION;
-        hashPrevBlock.SetNull();
-        hashMerkleRoot.SetNull();
-        nTime = 0;
-        nBits = 0;
-        nNonce = 0;
-        m_hash_cache.SetNull();
-    }
-
-    bool IsNull() const
-    {
-        return (nBits == 0);
-    }
-
-    uint256 GetHash(const bool use_cache = false) const
-    {
-        // The block hash cache field prevents repeated computations of the
-        // block's hash in the block acceptance pipeline. It's particularly
-        // effective for early blocks with expensive scrypt hashes. Dynamic
-        // caching isn't the prettiest solution, but it provides an interim
-        // performance advantage as we refactor legacy code.
-        //
-        // use_cache defaults to false to discourage use of the cache except
-        // in carefully chosen single-threaded scenarios. Avoid hash caching
-        // for block objects except where thread-safety is obvious and where
-        // performance improves significantly.
-        //
-        if (use_cache) {
-            if (!m_hash_cache.IsNull()) {
-                return m_hash_cache;
-            }
-
-            m_hash_cache = ComputeHash();
-
-            return m_hash_cache;
-        }
-
-        return ComputeHash();
-    }
-
-    uint256 GetPoWHash() const
-    {
-        return scrypt_blockhash(CVOIDBEGIN(nVersion));
-    }
-
-    int64_t GetBlockTime() const
-    {
-        return (int64_t)nTime;
-    }
-
-private:
-    mutable uint256 m_hash_cache;
-
-    uint256 ComputeHash() const
-    {
-        if (nVersion >= 7) {
-            return SerializeHash(*this);
-        }
-
-        return GetPoWHash();
-    }
-};
-
-namespace GRC {
-//!
-//! \brief A report that contains the mrc fees paid in a block.
-//!
-class MRCFees
-{
-public:
-    CAmount m_mrc_foundation_fees = 0;   //!< mrc fees to the foundation
-    CAmount m_mrc_staker_fees = 0;       //!< mrc fees to the staker
-    CAmount m_mrc_minimum_calc_fees = 0; //!< minimum calculated mrc fees for validation
-};
-
-//!
-//! \brief A report that contains the calculated subsidy claimed in a block.
-//! Produced by the CBlock::GetMint() method.
-//!
-class MintSummary
-{
-public:
-    CAmount m_total = 0; //!< Total value claimed by the block producer.
-    CAmount m_fees = 0;  //!< Fees paid for the block's transactions.
-};
-}
-
-class CBlock : public CBlockHeader
-{
-public:
-    // network and disk
-    std::vector<CTransaction> vtx;
-
-    // ppcoin: block signature - signed by one of the coin base txout[N]'s owner
-    std::vector<unsigned char> vchBlockSig;
-
-    // memory only
-    mutable bool fChecked;
-
-    //! Lazy cache for the parsed claim contract on legacy (pre-v11) blocks.
-    mutable GRC::Contract m_claim_contract_cache;
-
-    CBlock()
-    {
-        SetNull();
-    }
-
-    CBlock(const CBlockHeader &header)
-    {
-        SetNull();
-        *(static_cast<CBlockHeader*>(this)) = header;
-    }
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action)
-    {
-        READWRITEAS(CBlockHeader, *this);
-
-        // ConnectBlock depends on vtx following header to generate CDiskTxPos
-        if (!(s.GetType() & (SER_GETHASH|SER_BLOCKHEADERONLY))) {
-            READWRITE(vtx);
-            READWRITE(vchBlockSig);
-        } else if (ser_action.ForRead()) {
-            const_cast<CBlock*>(this)->vtx.clear();
-            const_cast<CBlock*>(this)->vchBlockSig.clear();
-        }
-    }
-
-    void SetNull()
-    {
-        CBlockHeader::SetNull();
-
-        vtx.clear();
-        vchBlockSig.clear();
-        fChecked = false;
-        m_claim_contract_cache = GRC::Contract();
-    }
-
-    CBlockHeader GetBlockHeader() const
-    {
-        CBlockHeader block;
-        block.nVersion       = nVersion;
-        block.hashPrevBlock  = hashPrevBlock;
-        block.hashMerkleRoot = hashMerkleRoot;
-        block.nTime          = nTime;
-        block.nBits          = nBits;
-        block.nNonce         = nNonce;
-        return block;
-    }
-
-    const GRC::Claim& GetClaim() const;
-    GRC::Claim PullClaim();
-    GRC::SuperblockPtr GetSuperblock() const;
-    GRC::SuperblockPtr GetSuperblock(const CBlockIndex* const pindex) const;
-    GRC::MintSummary GetMint() const;
-    GRC::MRCFees GetMRCFees() const;
-
-    // entropy bit for stake modifier if chosen by modifier
-    unsigned int GetStakeEntropyBit() const
-    {
-        // Take last bit of block hash as entropy bit
-        unsigned int nEntropyBit = ((GetHash(true).GetUint64(0)) & 1llu);
-        if (LogInstance().WillLogCategory(BCLog::LogFlags::VERBOSE) && gArgs.GetBoolArg("-printstakemodifier"))
-            LogPrintf("GetStakeEntropyBit: hashBlock=%s nEntropyBit=%u", GetHash(true).ToString(), nEntropyBit);
-        return nEntropyBit;
-    }
-
-    // ppcoin: two types of block: proof-of-work or proof-of-stake
-    bool IsProofOfStake() const
-    {
-        return (vtx.size() > 1 && vtx[1].IsCoinStake());
-    }
-
-    bool IsProofOfWork() const
-    {
-        return !IsProofOfStake();
-    }
-
-    // ppcoin: get max transaction timestamp
-    int64_t GetMaxTransactionTime() const
-    {
-        int64_t maxTransactionTime = 0;
-        for (auto const& tx : vtx)
-            maxTransactionTime = std::max(maxTransactionTime, (int64_t)tx.nTime);
-        return maxTransactionTime;
-    }
-
-
-    void print() const
-    {
-        LogPrintf("CBlock(hash=%s, ver=%d, hashPrevBlock=%s, hashMerkleRoot=%s, nTime=%u, nBits=%08x, nNonce=%u, vtx=%" PRIszu ", vchBlockSig=%s)",
-            GetHash().ToString(),
-            nVersion,
-            hashPrevBlock.ToString(),
-            hashMerkleRoot.ToString(),
-            nTime, nBits, nNonce,
-            vtx.size(),
-            HexStr(vchBlockSig));
-        for (unsigned int i = 0; i < vtx.size(); i++)
-        {
-            LogPrintf("  ");
-            vtx[i].print();
-        }
-    }
-};
-
-/** The block chain is a tree shaped structure starting with the
- * genesis block at the root, with each block potentially having multiple
- * candidates to be the next block.  pprev and pnext link a path through the
- * main/longest chain.  A blockindex may have multiple pprev pointing back
- * to it, but pnext will only point forward to the longest branch, or will
- * be null if the block is not part of the longest chain.
- */
-class CBlockIndex
-{
-public:
-    const uint256* phashBlock;
-    CBlockIndex* pprev;
-    CBlockIndex* pnext;
-    unsigned int nFile;
-    unsigned int nBlockPos;
-    int64_t nMoneySupply;
-    GRC::ResearcherContext* m_researcher;
-    std::vector<GRC::ResearcherContext*> m_mrc_researchers;
-    int nHeight;
-
-    unsigned int nFlags;  // ppcoin: block index flags
-    enum : uint32_t
-    {
-        BLOCK_PROOF_OF_STAKE = (1 << 0), // is proof-of-stake block
-        BLOCK_STAKE_ENTROPY  = (1 << 1), // entropy bit for stake modifier
-        BLOCK_STAKE_MODIFIER = (1 << 2), // regenerated stake modifier
-
-        // Gridcoin
-        EMPTY_CPID           = (1 << 3), // CPID is empty
-        NONCRUNCHER_CPID     = (1 << 4), // CPID equals "NONCRUNCHER"
-        SUPERBLOCK           = (1 << 5), // Block contains a superblock
-        CONTRACT             = (1 << 6), // Block contains a contract
-    };
-
-    uint64_t nStakeModifier; // hash modifier for proof-of-stake
-    uint256 hashProof;
-
-    // block header
-    int nVersion;
-    unsigned int nTime;
-    unsigned int nBits;
-    unsigned int nNonce;
-    uint256 hashMerkleRoot;
-
-    CBlockIndex()
-    {
-        SetNull();
-    }
-
-    CBlockIndex(unsigned int nFileIn, unsigned int nBlockPosIn, CBlock& block)
-    {
-        SetNull();
-
-        nFile = nFileIn;
-        nBlockPos = nBlockPosIn;
-        if (block.IsProofOfStake())
-        {
-            SetProofOfStake();
-        }
-
-        nVersion       = block.nVersion;
-        hashMerkleRoot = block.hashMerkleRoot;
-        nTime          = block.nTime;
-        nBits          = block.nBits;
-        nNonce         = block.nNonce;
-    }
-
-    void SetNull()
-    {
-        phashBlock = nullptr;
-        pprev = nullptr;
-        pnext = nullptr;
-        nFile = 0;
-        nBlockPos = 0;
-        nHeight = 0;
-        nMoneySupply = 0;
-        nFlags = EMPTY_CPID;
-        nStakeModifier = 0;
-        hashProof.SetNull();
-
-        nVersion       = 0;
-        hashMerkleRoot.SetNull();
-        nTime          = 0;
-        nBits          = 0;
-        nNonce         = 0;
-
-        // Note that the pointer entries are deleted, but not the memory allocations in the pool to which they
-        // were allocated. I think this means that pool entries are leaked (orphaned) under reorgs.
-        // TODO: Investigate this carefully as it could be a small memory leak source.
-        m_researcher = nullptr;
-        m_mrc_researchers.clear();
-    }
-
-    CBlockHeader GetBlockHeader() const
-    {
-        CBlockHeader block;
-        block.nVersion       = nVersion;
-        if (pprev)
-            block.hashPrevBlock = pprev->GetBlockHash();
-        block.hashMerkleRoot = hashMerkleRoot;
-        block.nTime          = nTime;
-        block.nBits          = nBits;
-        block.nNonce         = nNonce;
-        return block;
-    }
-
-    uint256 GetBlockHash() const
-    {
-        return *phashBlock;
-    }
-
-    int64_t GetBlockTime() const
-    {
-        return (int64_t)nTime;
-    }
-
-    arith_uint256 GetBlockTrust() const;
-
-    bool IsInMainChain() const EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-    {
-        return (pnext || this == pindexBest);
-    }
-
-    int64_t GetPastTimeLimit() const
-    {
-        return GetMedianTimePast();
-    }
-
-    enum { nMedianTimeSpan=11 };
-
-    int64_t GetMedianTimePast() const
-    {
-        int64_t pmedian[nMedianTimeSpan];
-        int64_t* pbegin = &pmedian[nMedianTimeSpan];
-        int64_t* pend = &pmedian[nMedianTimeSpan];
-
-        const CBlockIndex* pindex = this;
-        for (int i = 0; i < nMedianTimeSpan && pindex; i++, pindex = pindex->pprev)
-            *(--pbegin) = pindex->GetBlockTime();
-
-        std::sort(pbegin, pend);
-        return pbegin[(pend - pbegin)/2];
-    }
-
-    bool IsProofOfWork() const
-    {
-        return !(nFlags & BLOCK_PROOF_OF_STAKE);
-    }
-
-    bool IsProofOfStake() const
-    {
-        return (nFlags & BLOCK_PROOF_OF_STAKE);
-    }
-
-    void SetProofOfStake()
-    {
-        nFlags |= BLOCK_PROOF_OF_STAKE;
-    }
-
-    bool IsUserCPID() const
-    {
-        return !(nFlags & (NONCRUNCHER_CPID | EMPTY_CPID));
-    }
-
-    unsigned int GetStakeEntropyBit() const
-    {
-        return ((nFlags & BLOCK_STAKE_ENTROPY) >> 1);
-    }
-
-    bool SetStakeEntropyBit(unsigned int nEntropyBit)
-    {
-        if (nEntropyBit > 1)
-            return false;
-        nFlags |= (nEntropyBit ? BLOCK_STAKE_ENTROPY : (uint32_t) 0);
-        return true;
-    }
-
-    bool GeneratedStakeModifier() const
-    {
-        return (nFlags & BLOCK_STAKE_MODIFIER);
-    }
-
-    void SetStakeModifier(uint64_t nModifier, bool fGeneratedStakeModifier)
-    {
-        nStakeModifier = nModifier;
-        if (fGeneratedStakeModifier)
-            nFlags |= BLOCK_STAKE_MODIFIER;
-    }
-
-    void SetResearcherContext(
-        const GRC::MiningId mining_id,
-        const int64_t research_subsidy,
-        const double magnitude)
-    {
-        nFlags &= ~(EMPTY_CPID | NONCRUNCHER_CPID);
-
-        if (const auto cpid_option = mining_id.TryCpid()) {
-            if (research_subsidy > 0) {
-                if (!m_researcher) {
-                    m_researcher = GRC::BlockIndexPool::GetNextResearcherContext();
-                }
-
-                m_researcher->m_cpid = *cpid_option;
-                m_researcher->m_research_subsidy = research_subsidy;
-                m_researcher->m_magnitude = magnitude;
-
-                return;
-            }
-        }
-
-        if (m_researcher) {
-            delete m_researcher;
-            m_researcher = nullptr;
-        }
-
-        if (mining_id.Which() == GRC::MiningId::Kind::INVALID) {
-            nFlags |= EMPTY_CPID;
-        } else {
-            nFlags |= NONCRUNCHER_CPID;
-        }
-    }
-
-    void AddMRCResearcherContext(
-            const GRC::MiningId mining_id,
-            const int64_t research_subsidy,
-            const double magnitude)
-    {
-        if (const GRC::CpidOption cpid_option = mining_id.TryCpid()) {
-            if (research_subsidy > 0) {
-                // Check for duplicate insert (the m_mrc_researchers is a vector not a map).
-                for (const auto& iter : m_mrc_researchers) {
-                    if (iter->m_cpid == *cpid_option) {
-                        LogPrintf("WARNING: %s: Ignoring duplicate AddMRCResearcherContext for cpid %s, research_subsidy "
-                                  "%s",
-                                  __func__,
-                                  cpid_option->ToString(),
-                                  FormatMoney(iter->m_research_subsidy)
-                                  );
-
-                        return;
-                    }
-                }
-
-                GRC::ResearcherContext* mrc_researcher = GRC::BlockIndexPool::GetNextResearcherContext();
-
-                mrc_researcher->m_cpid = *cpid_option;
-                mrc_researcher->m_research_subsidy = research_subsidy;
-                mrc_researcher->m_magnitude = magnitude;
-
-                m_mrc_researchers.push_back(mrc_researcher);
-
-                return;
-            }
-        }
-    }
-
-    GRC::MiningId GetMiningId() const
-    {
-        if (m_researcher)
-            return GRC::MiningId(m_researcher->m_cpid);
-
-        if (nFlags & NONCRUNCHER_CPID)
-            return GRC::MiningId::ForNoncruncher();
-
-        return GRC::MiningId();
-    }
-
-    int64_t ResearchSubsidy() const
-    {
-        if (m_researcher) {
-            return m_researcher->m_research_subsidy;
-        }
-
-        return 0;
-    }
-
-    int64_t ResearchMRCSubsidy() const
-    {
-        int64_t mrc_subsidy = 0;
-
-        for (const auto& mrc_researcher : m_mrc_researchers) {
-            mrc_subsidy += mrc_researcher->m_research_subsidy;
-        }
-
-        return mrc_subsidy;
-    }
-
-    double Magnitude() const
-    {
-        if (m_researcher) {
-            return m_researcher->m_magnitude;
-        }
-
-        return 0;
-    }
-
-    bool IsSuperblock() const
-    {
-        return nFlags & SUPERBLOCK;
-    }
-
-    void MarkAsSuperblock()
-    {
-        nFlags |= SUPERBLOCK;
-    }
-
-    bool IsContract() const
-    {
-        return nFlags & CONTRACT;
-    }
-
-    void MarkAsContract()
-    {
-        nFlags |= CONTRACT;
-    }
-
-    std::string ToString() const
-    {
-        return strprintf("CBlockIndex(nprev=%p, pnext=%p, nFile=%u, nBlockPos=%-6d nHeight=%d, nMoneySupply=%s, nFlags=(%s)(%d)(%s), nStakeModifier=%016" PRIx64 ", hashProof=%s, merkle=%s, hashBlock=%s)",
-            pprev, pnext, nFile, nBlockPos, nHeight,
-            FormatMoney(nMoneySupply),
-            GeneratedStakeModifier() ? "MOD" : "-", GetStakeEntropyBit(), IsProofOfStake()? "PoS" : "PoW",
-            nStakeModifier,
-            hashProof.ToString(),
-            hashMerkleRoot.ToString(),
-            GetBlockHash().ToString());
-    }
-
-    void print() const
-    {
-        LogPrintf("%s", ToString());
-    }
-};
-
-
-
-/** Used to marshal pointers into hashes for db storage. */
-class CDiskBlockIndex : public CBlockIndex
-{
-private:
-    uint256 blockHash;
-
-public:
-    uint256 hashPrev;
-    uint256 hashNext;
-
-    CDiskBlockIndex()
-    {
-        hashPrev.SetNull();
-        hashNext.SetNull();
-        blockHash.SetNull();
-    }
-
-    explicit CDiskBlockIndex(CBlockIndex* pindex) : CBlockIndex(*pindex)
-    {
-        hashPrev = (pprev ? pprev->GetBlockHash() : uint256());
-        hashNext = (pnext ? pnext->GetBlockHash() : uint256());
-    }
-
-    void AddMRCResearcherContextFromDisk(const GRC::ResearcherContext& mrc_disk)
-    {
-        GRC::ResearcherContext* mrc_researcher = GRC::BlockIndexPool::GetNextResearcherContext();
-
-        mrc_researcher->m_cpid = mrc_disk.m_cpid;
-        mrc_researcher->m_research_subsidy = mrc_disk.m_research_subsidy;
-        mrc_researcher->m_magnitude = mrc_disk.m_magnitude;
-
-        m_mrc_researchers.push_back(mrc_researcher);
-    }
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action)
-    {
-        if (!(s.GetType() & SER_GETHASH)) {
-            READWRITE(nVersion);
-        }
-
-        READWRITE(hashNext);
-        READWRITE(nFile);
-        READWRITE(nBlockPos);
-        READWRITE(nHeight);
-        int64_t nMint = 0; // removed
-        READWRITE(nMint);
-        READWRITE(nMoneySupply);
-        READWRITE(nFlags);
-        READWRITE(nStakeModifier);
-
-        // Duplicate stake detection removed from the block index:
-        if (IsProofOfStake()) {
-            COutPoint prevoutStake;
-            uint32_t nStakeTime = 0;
-
-            READWRITE(prevoutStake);
-            READWRITE(nStakeTime);
-        }
-
-        READWRITE(hashProof);
-
-        // block header
-        READWRITE(nVersion);
-        READWRITE(hashPrev);
-        READWRITE(hashMerkleRoot);
-        READWRITE(nTime);
-        READWRITE(nBits);
-        READWRITE(nNonce);
-        READWRITE(blockHash);
-
-        //7-11-2015 - Gridcoin - New Accrual Fields (Note, Removing the deterministic block number to make this happen all the time):
-        std::string cpid_hex = GetMiningId().ToString();
-        double research_subsidy_grc = ResearchSubsidy() / (double)COIN;
-        double interest_subsidy_grc = 0; // removed
-        double magnitude = Magnitude();
-
-        READWRITE(cpid_hex);
-        READWRITE(research_subsidy_grc);
-        READWRITE(interest_subsidy_grc);
-        READWRITE(magnitude);
-
-        // Superblock and contract flags merged into nFlags:
-        uint32_t is_superblock = this->IsSuperblock();
-        uint32_t is_contract = this->IsContract();
-
-        if (IsResearchAgeEnabled(this->nHeight)) {
-            READWRITE(is_superblock);
-            READWRITE(is_contract);
-
-            std::string dummy;
-
-            // Blocks used to contain the GRC address.
-            READWRITE(dummy);
-
-            // Blocks used to come with a reserved string. Keep (de)serializing
-            // it until it's used.
-            READWRITE(dummy);
-        }
-
-        // Translate legacy disk format to memory format:
-        //
-        if (ser_action.ForRead()) {
-            NCONST_PTR(this)->SetResearcherContext(
-                GRC::MiningId::Parse(cpid_hex),
-                research_subsidy_grc * COIN,
-                magnitude);
-
-            if (is_superblock == 1) {
-                NCONST_PTR(this)->MarkAsSuperblock();
-            }
-
-            if (is_contract == 1) {
-                NCONST_PTR(this)->MarkAsContract();
-            }
-        }
-
-        if (nVersion >= 12) {
-            std::vector<GRC::ResearcherContext> m_mrc_researchers_disk;
-
-            for (const auto& mrc : m_mrc_researchers) {
-                m_mrc_researchers_disk.push_back(*mrc);
-            }
-
-            READWRITE(m_mrc_researchers_disk);
-
-            if (ser_action.ForRead()) {
-                for (const auto& mrc : m_mrc_researchers_disk) {
-                    NCONST_PTR(this)->AddMRCResearcherContextFromDisk(mrc);
-                }
-            }
-        }
-    }
-
-    uint256 GetBlockHash() const
-    {
-        if (fUseFastIndex && (nTime < GetAdjustedTime() - 24 * 60 * 60) && !blockHash.IsNull())
-            return blockHash;
-
-        CBlockHeader block;
-        block.nVersion        = nVersion;
-        block.hashPrevBlock   = hashPrev;
-        block.hashMerkleRoot  = hashMerkleRoot;
-        block.nTime           = nTime;
-        block.nBits           = nBits;
-        block.nNonce          = nNonce;
-
-        const_cast<CDiskBlockIndex*>(this)->blockHash = block.GetHash();
-
-        return blockHash;
-    }
-
-    std::string ToString() const
-    {
-        std::string str = "CDiskBlockIndex(";
-        str += CBlockIndex::ToString();
-        str += strprintf("\n                hashBlock=%s, hashPrev=%s, hashNext=%s)",
-            GetBlockHash().ToString(),
-            hashPrev.ToString(),
-            hashNext.ToString());
-        return str;
-    }
-
-    void print() const
-    {
-        LogPrintf("%s", ToString());
-    }
-};
-
-
-
-
-
-/** Describes a place in the block chain to another node such that if the
- * other node doesn't have the same branch, it can find a recent common trunk.
- * The further back it is, the further before the fork it may be.
- */
-class CBlockLocator
-{
-public:
-    std::vector<uint256> vHave;
-
-    CBlockLocator()
-    {
-    }
-
-    explicit CBlockLocator(const CBlockIndex* pindex)
-    {
-        Set(pindex);
-    }
-
-    explicit CBlockLocator(uint256 hashBlock)
-    {
-        BlockMap::iterator mi = mapBlockIndex.find(hashBlock);
-        if (mi != mapBlockIndex.end())
-            Set(mi->second);
-    }
-
-    CBlockLocator(const std::vector<uint256>& vHaveIn)
-    {
-        vHave = vHaveIn;
-    }
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action)
-    {
-        if (!(s.GetType() & SER_GETHASH)) {
-            int nVersion = s.GetVersion();
-            READWRITE(nVersion);
-        }
-
-        READWRITE(vHave);
-    }
-
-    void SetNull()
-    {
-        vHave.clear();
-    }
-
-    bool IsNull()
-    {
-        return vHave.empty();
-    }
-
-    void Set(const CBlockIndex* pindex)
-    {
-        vHave.clear();
-        int nStep = 1;
-        while (pindex)
-        {
-            vHave.push_back(pindex->GetBlockHash());
-
-            // Exponentially larger steps back
-            for (int i = 0; pindex && i < nStep; i++)
-                pindex = pindex->pprev;
-            if (vHave.size() > 10)
-                nStep *= 2;
-        }
-        vHave.push_back((Params().IsMockableChain() ? hashGenesisBlockRegTest : !fTestNet ? hashGenesisBlock : hashGenesisBlockTestNet));
-    }
-
-    int GetDistanceBack() EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-    {
-        // Retrace how far back it was in the sender's branch
-        int nDistance = 0;
-        int nStep = 1;
-        for (auto const& hash : vHave)
-        {
-            BlockMap::iterator mi = mapBlockIndex.find(hash);
-            if (mi != mapBlockIndex.end())
-            {
-                CBlockIndex* pindex = mi->second;
-                if (pindex->IsInMainChain())
-                    return nDistance;
-            }
-            nDistance += nStep;
-            if (nDistance > 10)
-                nStep *= 2;
-        }
-        return nDistance;
-    }
-
-    CBlockIndex* GetBlockIndex() EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-    {
-        // Find the first block the caller has in the main chain
-        for (auto const& hash : vHave)
-        {
-            BlockMap::iterator mi = mapBlockIndex.find(hash);
-            if (mi != mapBlockIndex.end())
-            {
-                CBlockIndex* pindex = mi->second;
-                if (pindex->IsInMainChain())
-                    return pindex;
-            }
-        }
-        return pindexGenesisBlock;
-    }
-
-    uint256 GetBlockHash() EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-    {
-        // Find the first block the caller has in the main chain
-        for (auto const& hash : vHave)
-        {
-            BlockMap::iterator mi = mapBlockIndex.find(hash);
-            if (mi != mapBlockIndex.end())
-            {
-                CBlockIndex* pindex = mi->second;
-                if (pindex->IsInMainChain())
-                    return hash;
-            }
-        }
-        return (Params().IsMockableChain() ? hashGenesisBlockRegTest : !fTestNet ? hashGenesisBlock : hashGenesisBlockTestNet);
-    }
-
-    int GetHeight() EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-    {
-        CBlockIndex* pindex = GetBlockIndex();
-        if (!pindex)
-            return 0;
-        return pindex->nHeight;
-    }
-};
